# coalgov-ai/docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "CoalGov AI API",
    "description": "Mine governance pilot: inspections, compliance, production approvals, grievances and offline field reporting.",
    "version": "1.1.0"
  },
  "paths": {
...
    "/api/telemetry": {
      "post": {
        "tags": [
          "Intelligence and reporting"
        ],
        "summary": "Telemetry",
        "operationId": "telemetry_api_telemetry_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SensorInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
    "/api/integrations": {
      "get": {
        "tags": [
          "Integration adapters"
        ],
        "summary": "Status",
        "operationId": "status_api_integrations_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/integrations/vision-events": {
      "post": {
        "tags": [
          "Integration adapters"
        ],
        "summary": "Vision Event",
        "operationId": "vision_event_api_integrations_vision_events_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VisionEvent"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "SensorInput": {
        "properties": {
          "mine_id": {
            "type": "string",
            "title": "Mine Id"
          },
          "asset_id": {
            "type": "string",
            "title": "Asset Id"
          },
          "source": {
            "type": "string",
            "maxLength": 80,
            "minLength": 2,
            "title": "Source"
          },
          "event_id": {
            "type": "string",
            "maxLength": 120,
            "minLength": 5,
            "title": "Event Id"
          },
          "metric": {
            "type": "string",
            "enum": [
              "methane",
              "temperature",
              "dust",
              "vibration"
            ],
            "title": "Metric"
          },
          "value": {
            "type": "number",
            "maximum": 100000.0,
            "minimum": 0.0,
            "title": "Value"
          },
          "unit": {
            "type": "string",
            "maxLength": 30,
            "minLength": 1,
            "title": "Unit"
          },
          "observed_at": {
            "type": "string",
            "format": "date-time",
            "title": "Observed At"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "mine_id",
          "asset_id",
          "source",
          "event_id",
          "metric",
          "value",
          "unit",
          "observed_at"
        ],
        "title": "SensorInput"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
      "VisionEvent": {
        "properties": {
          "event_id": {
            "type": "string",
            "maxLength": 64,
            "minLength": 8,
            "title": "Event Id"
          },
          "mine_id": {
            "type": "string",
            "title": "Mine Id"
          },
          "asset_id": {
            "type": "string",
            "title": "Asset Id"
          },
          "detection": {
            "type": "string",
            "enum": [
              "missing_ppe",
              "restricted_zone_entry",
              "smoke",
              "vehicle_proximity"
            ],
            "title": "Detection"
          },
          "confidence": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Confidence"
          },
          "observed_at": {
            "type": "string",
            "format": "date-time",
            "title": "Observed At"
          },
          "description": {
            "type": "string",
            "maxLength": 2000,
            "minLength": 10,
            "title": "Description"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "event_id",
          "mine_id",
          "asset_id",
          "detection",
          "confidence",
          "observed_at",
          "description"
        ],
        "title": "VisionEvent"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}